refactor: extract ManualIssueExtractionStatusPresenter.swift from IssueExtractionPresenters.swift
Splits the `ManualIssueExtractionStatusPresenter` @MainActor class
(~57 lines) out of IssueExtractionPresenters.swift into its own file.
IssueExtractionPresenters.swift retains the IssueExtractionStatusPresenter
enum + IssueExtractionFailurePresenter class.

Byte-identical move of lines 13-68. Both files import Foundation.

IssueExtractionPresenters.swift: 87 → 29 lines. Tests: 667.
